claims에 Audience로 cameraId를 넘겨주어 요청한 cameraId와 Audience가 다르면 서비스 거절

+++ src/main/java/com/example/falldowndetectionserver/controller/UserController.java
@@ -1,9 +1,11 @@
 package com.example.falldowndetectionserver.controller;
 
 import com.example.falldowndetectionserver.domain.dto.UserDTO;
+import com.example.falldowndetectionserver.jwt.TokenProvider;
 import com.example.falldowndetectionserver.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.catalina.User;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -19,6 +21,7 @@ import java.nio.charset.StandardCharsets;
 @Slf4j
 public class UserController {
     private final UserService userService;
+    private final TokenProvider tokenProvider;
 
     /**
      * 회원가입
@@ -37,13 +40,20 @@ public class UserController {
      */
     @GetMapping("info")
     @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
-    public ResponseEntity<UserDTO> getUserInfo(String cameraId) {
-        UserDTO userDTO = userService.getUserInfo(cameraId);
-        if (userDTO == null) {
-
-        }
+    public ResponseEntity<UserDTO> getUserInfo(String cameraId, @RequestHeader("Authorization") String token) {
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.setContentType(new MediaType("application", "json", StandardCharsets.UTF_8));
+
+        UserDTO userDTO;
+
+        if (!tokenProvider.getAudience(token.substring(7)).equals(cameraId)) {
+            userDTO = UserDTO.builder()
+                    .requestSuccess("NOT ALLOWED")
+                    .build();
+            return new ResponseEntity<>(userDTO, httpHeaders, HttpStatus.OK);
+        }
+
+        userDTO = userService.getUserInfo(cameraId);
 
         return new ResponseEntity<>(userDTO, httpHeaders, HttpStatus.OK);
     }
@@ -54,7 +64,10 @@ public class UserController {
      */
     @DeleteMapping("remove")
     @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
-    public ResponseEntity<String> removeUserInfo(String cameraId) {
+    public ResponseEntity<String> removeUserInfo(String cameraId, @RequestHeader("Authorization") String token) {
+        if (!tokenProvider.getAudience(token.substring(7)).equals(cameraId)) {
+            return ResponseEntity.ok("NOT ALLOWED");
+        }
         return ResponseEntity.ok(userService.removeUserInfo(cameraId));
     }
 
@@ -64,7 +77,10 @@ public class UserController {
      */
     @PutMapping("modify")
     @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
-    public ResponseEntity<String> modifyUserInfo(@RequestBody UserDTO userDTO) {
+    public ResponseEntity<String> modifyUserInfo(@RequestBody UserDTO userDTO, @RequestHeader("Authorization") String token) {
+        if (!tokenProvider.getAudience(token.substring(7)).equals(userDTO.getCameraId())) {
+            return ResponseEntity.ok("NOT ALLOWED");
+        }
         return ResponseEntity.ok(userService.modifyUserInfo(userDTO));
     }
 }
